make directory routers
