+++ app/main.py
@@ -1,0 +1,5 @@
+from fastapi import FastAPI
+
+from .routers import welcome
+app = FastAPI()
+
